doc-bot-flask/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
autopep8==2.3.1
azure-core==1.31.0
azure-storage-blob==12.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.8.0.post1
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.125
MarkupSafe==2.1.5
marshmallow==3.22.0
msal==1.31.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.47.0
orjson==3.10.7
packaging==24.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypdf==5.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
